Restructure as monorepo with query builder package
- Consolidate from monorepo to single package structure
- Rename package to "sparql-builder" for better discoverability
- Merge query-builder and client into unified src/ directory
- Update build configuration to use vite-plugin-dts
- Fix dynamic import warning by using static imports
- All 23 tests passing with full type safety
