--- packages/client/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@hideokamoto/simple-sparql-client",
-  "version": "1.0.1",
-  "description": "Simple Sparql Client",
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "files": ["dist"],
-  "scripts": {
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "build": "vite build",
-    "lint": "biome check .",
-    "lint:fix": "biome check --write .",
-    "format": "biome format --write .",
-    "prepublishOnly": "npm run build"
-  },
-  "keywords": ["sparql", "fetch", "semantic-web", "rdf", "typescript"],
-  "bugs": {
-    "url": "https://github.com/hideokamoto/sparql-client/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/hideokamoto/sparql-client.git"
-  },
-  "author": "Hidetaka Okamoto <info@wp-kyoto.net> (https://wp-kyoto.net)",
-  "license": "MIT",
-  "dependencies": {
-    "@hideokamoto/sparql-query-builder": "*",
-    "sparql-http-client": "^2.4.1"
-  },
-  "devDependencies": {
-    "@types/sparql-http-client": "^3.0.5"
-  }
-}
