--- packages/query-builder/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@hideokamoto/sparql-query-builder",
-  "version": "1.0.0",
-  "description": "Type-safe SPARQL query builder for JavaScript/TypeScript",
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "files": ["dist"],
-  "scripts": {
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "build": "vite build",
-    "lint": "biome check .",
-    "lint:fix": "biome check --write .",
-    "format": "biome format --write .",
-    "prepublishOnly": "npm run build"
-  },
-  "keywords": ["sparql", "query-builder", "rdf", "semantic-web", "typescript"],
-  "bugs": {
-    "url": "https://github.com/hideokamoto/sparql-client/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/hideokamoto/sparql-client.git",
-    "directory": "packages/query-builder"
-  },
-  "author": "Hidetaka Okamoto <info@wp-kyoto.net> (https://wp-kyoto.net)",
-  "license": "MIT"
-}
