+++ src/index.ts
@@ -1,11 +1,20 @@
-import { type QueryType, SPARQLClient, type SparqlBinding, type SparqlResults } from './client.js'
+// Re-export QueryBuilder
+export { QueryBuilder, type OrderDirection, type Prefix, type Triple } from './query-builder.js'
 
-// Re-export QueryBuilder from query-builder package
-export { QueryBuilder, type OrderDirection } from '@hideokamoto/sparql-query-builder'
+// Re-export Client
+export {
+  SPARQLClient,
+  type QueryType,
+  type SparqlBinding,
+  type SparqlResults,
+} from './client.js'
+
+// Import for local use
+import { SPARQLClient, type SparqlBinding } from './client.js'
 
 /**
  * Execute SPARQL query as a simple function
- * @param query - SPARQL query string or QueryBuilder instance
+ * @param query - SPARQL query string
  * @param endpoint - SPARQL endpoint URL (default: DBpedia)
  * @returns Promise with query result bindings
  */
@@ -17,5 +26,3 @@ export const execSparqlQuery = async (
   client.setQuery(query)
   return client.get()
 }
-
-export { SPARQLClient, type QueryType, type SparqlBinding, type SparqlResults }
